pkgs/profpatsch/execline: change el_semicolon to only read one arg
The C implementation of el_semicolon in execline only reads one
argument at a time and returns an index into the rest of argv.

This makes sense for the usual application of programs using it, which
is just reading a few arguments and a block or two, and then executing
into `prog`. `prog` could be anything really, including additional
blocks.

The new `el_semicolon_full_argv` function exports the previous
behaviour of parsing the whole thing.

As a nice side-effect, we return the rest of argv in-place.