pkgs/profpatsch/xdg-open: Prepare decoding http urls
I want to be able to open http(s) links that are e.g. images directly
in the right application. Aka web urls should be transparent, instead
of always opening everthing in the browser.

This adds some silly ways of connecting to the server and parsing
out the headers, in order to fetch the content-type.