pkgs: Update easy-dhall-nix to latest master
I'm not familiar with the "easy-dhall-nix" project, but the repository
is imported into Vuizvui via import-from-derivation. While this by
itself is not a big issue (apart from contributing hugely to evaluation
time, we're already at around an hour), the "dhall-nix" derivation is in
turn imported again via importDhall, so whenever something breaks with
dhall-nix, our evaluation will break as well.

Unfortunately, something is broken right now:

  building '/nix/store/c363947v9qk287d07qj2kpj60rmzwalj-dhall-nix-1.1.14-x86_64-linux.tar.bz2.drv'...

  trying https://github.com/dhall-lang/dhall-haskell/releases/download/1.32.0/dhall-nix-1.1.14-x86_64-linux.tar.bz2
    % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                   Dload  Upload   Total   Spent    Left  Speed
  100   648  100   648    0     0   2592      0 --:--:-- --:--:-- --:--:--  2581
  100 2255k  100 2255k    0     0  1639k      0  0:00:01  0:00:01 --:--:-- 6287k
  hash mismatch in fixed-output derivation '/nix/store/yhls1ffnvp1nbjsm0xr3l1z6j6x4waqh-dhall-nix-1.1.14-x86_64-linux.tar.bz2':
    wanted: sha256:1j32jf0is0kikfw7h9w3n8ikw70bargr32d1cyasqgmb7s7mvs1c
    got:    sha256:1qs5p05qfk5xs1ajwyhn27m0bzs96lnlf3b4gnkffajhaq7hiqll
  cannot build derivation '/nix/store/aj5ag721b9gm4an6yxh2ljg19ixg4alv-dhall-nix-simple.drv': 1 dependencies couldn't be built

The reason why this happens is because GitHub's tarballs are not
deterministic and whenever GitHub changes something in the way these are
generated, we get a hash mismatch.

For exactly that reason, the fetchFromGitHub wrapper in nixpkgs uses
fetchzip instead of fetchurl, so that file ordering in the archive
doesn't matter.

Unfortunately, the upstream project still uses fetchurl, but since the
URLs and hashes have changed due to a bump to Dhall version 1.33.1, I've
choosen to switch to latest master instead of monkeypatching via
extraPostFetch.

With this bandaid, we shouldn't run into a hash collision until either
the next GC or until the upstream project has switched to either
fetchFromGitHub or fetchzip.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch, @justinwoo