workstation: Disable pull.rebase in Git config
From the release notes of Git version 2.27.0:

 * "git pull" issues a warning message until the pull.rebase
   configuration variable is explicitly given, which some existing
   users may find annoying---those who prefer not to rebase need to
   set the variable to false to squelch the warning.

This is exactly the warning which is annoying me all the time now, so in
order to get rid of it, let's explicitly set the default behaviour
(which is doing a recursive merge).

Just to be sure that I really want the default behaviour, I analysed my
shell history for invocations of "git pull" and only around 20% of the
invocations were with --rebase, 14% were with an explicit URL (but no
rebase) and the rest were recursive merges.

Signed-off-by: aszlig <aszlig@nix.build>