dnyarri: Add youtube-dl support to gPodder
I'm using gPodder to follow a bunch of YouTube channels and the internal
extractor/downloader tends to be pretty unreliable.

On the other hand, youtube-dl is regularily updated and supports a ton
of different formats.

To make sure we can actually use the extension, gPodder needs to have
access to the youtube_dl Python module, so we need to add it to the
propagatedBuildInputs.

Signed-off-by: aszlig <aszlig@nix.build>