pkgs/psi: Increment date in version number
I really don't have a lot of other things to blame for forgetting this
in the first place, but let's just assume that the temperatures here
right now are higher than I'm used[TM] to and thus my brain wasn't
working.

On the other hand while writing this I also took a peek and aparently
it's around 20 degree celsius right now, which would be... well...

... okay, I just forgot about that and I feel ashamed now. Go on!
Nothing to read here!

Signed-off-by: aszlig <aszlig@nix.build>