pkgs/aszlig: Add Mutt configuration
I've been using this configuration since years already but so far it has
been residing in ~/.muttrc and I copied to new machines accordingly.

The reason why I didn't add it here was because the config was too ugly
and I never got so far as to properly re-do it.

Unfortunately, the config is still ugly as hell, but at least we now
generate it from a structured Nix format and also the IMAP/SMTP user and
server infos are now retrieved via gopass instead.

This also includes my small prank multipart/alternative filter, which
should hopefully "encourange" recipients to disable HTML
parsing/rendering.

Signed-off-by: aszlig <aszlig@nix.build>