modules/profpatsch/weechat: add a workaround to attach ssh directly
So far we directly start a mosh-daemon when the user connects.
This breaks down in situations where UDP is blocked (e.g. some
hotspots). In that case, ssh can be used directly:

Example:

    ssh -t weechat@legosi ssh

The ssh argument tells it to connect directly.
Note the `-t`, which forces a pseudo-tty, otherwise tmux will
complain that it can’t find a terminal.