tests: Switch all tests to Python test driver
Since the removal[1] of the Perl test driver, our tests will no longer
run or even evaluate.

Fortunately, the test API is more or less the same, so the transition to
Python was not very involved.

However, I did add a "# fmt: off" on top of every testScript, since
formatting with black not only has issues with parameterised
antiquotations but is also plain ugly to mix 2 spaces of indentation
with 4 spaces of indentation.

Additionally, I'd like to have a maximum line length of 79 characters in
my Nix expressions while black on the other side even *insists* of using
longer lines.

[1]: https://github.com/NixOS/nixpkgs/commit/0620184f3f94f1bf8de014ab168

Signed-off-by: aszlig <aszlig@nix.build>