machines/haku: set NIX_PATH to filtered pkgs.path
This is a step towards a saner deployment for these machines.
The NIX_PATH is fixed, so that everything on the running system always
uses exactly the nixpkgs version the system was deployed with.
The deployment is done by copying the system closure and switching to
it via the `bin/switch-to-configuration` script.

Uses the reference to pkgs.path, and applies filterSourceGitignore on
the directory, since I often deploy from a local checkout which is
half a GB without the gitignore filter.