modules/gnupg: Remove "with lib;" on whole module
Using "with lib;" over such a broad scope makes it more difficult to
detect early evaluation errors since we need to evaluate the whole
system to check whether a non-existing attribute set is actually using
the "lib" fallback or whether it really exists in the current scope.

This makes "nix-instantiate --parse" feasible for detecting typos early
on.

Signed-off-by: aszlig <aszlig@nix.build>