Add basic flake.nix
Unfortunately, most of the functionality can't be easily exposed via Nix
Flakes, so this will be a very limited Flake with only our NixOS modules
and our packages as an overlay.

One of those things that are very hard to expose are our packaged games,
since the availability of them are dependent on the configuration (eg.
GOG, itch.io or HIB credentials).

Even when it comes to buildSandbox - which is the main reason why I'm
adding a flake.nix - I am not so sure what's the best way to expose it.

For example on one side, this could be exposed as
vuizvui.lib.buildSandbox, which in turn expects a pkgs argument, but on
the other side, we want to have certain other parts that *do* depend on
a locked version of nixpkgs. Using pkgs as part of a "lib" attribute
also sounds a bit weird to me, since lib usually doesn't contain
anything depending on stdenv.

So until we have a better way of integration, I'll just expose the
overlay and our NixOS modules.

Signed-off-by: aszlig <aszlig@nix.build>