devhell/packages: Remove gstreamer package
The package attributes for GSstreamer 0.x were dropped[1] a while ago,
so evaluating the packages no longer succeeds and instead we get an
error like this on Hydra:

  error: undefined variable 'gstreamer' at .../profiles/packages.nix:129:7

Initially I was tempted to change this to use pkgs.gst_all_1.gstreamer,
but looking at the differences between the old pkgs.gstreamer, the new
package only contains the relevant libraries while the old one contains
a few binaries, namely gst-feedback, gst-inspect, gst-launch,
gst-typefind, gst-xmlinspect, gst-xmllaunch and all of them again with a
"-0.10" suffix.

All of these tools are development tools and should not be used in real
applications, for example from the manual[2] on gst-launch-1.0:

> Please note that gst-launch-1.0 is primarily a debugging tool. You
> should not build applications on top of it. For applications, use the
> gst_parse_launch() function of the GStreamer API as an easy way to
> construct pipelines from pipeline descriptions.

While environment.systemPackages does link other directories than just
$out/bin, the primary reason why you'd want to put something in there is
to make it available in the system's PATH.

When asking @devhell about this he didn't remember the exact reason why
he put gstreamer in systemPackages, so I can only *assume* it was
because of tools like gst-launch, which unfortunately even I can
remember seeing recommended (instead of eg. ffmpeg) in some dark corners
of the web.

So if this would be really about gst-launch, we would need to put in
pkgs.gst_all_1.gstreamer.dev into systemPackages, since the development
tools are now where they belong.

Given that @devhell doesn't know the reason anymore and *also* told me
to remove it *and* it's also just a development tool which is not
supposed to be used in production, I'm hereby removing the package.

Should there really be applications which rely on this, they should
hopefully break after this change so they can be fixed to not rely on
these development tools.

[1]: https://github.com/NixOS/nixpkgs/commit/4a4e642abaaa026b55f42248a7b
[2]: https://gstreamer.freedesktop.org/documentation/tools/gst-launch.html

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @devhell