aszlig/gopass: Downgrade to version 1.9.2
Version 1.10.X switches the default to new MIME-based format which
breaks interoperability with pass and other implementations.

I'd gladly switch to that format, but the 1.10.0 changelog also
announces plans to remove support for GnuPG and Git in the long term:

> The goal is to remove the support for multiple backends and any
> external dependencies, including git and gpg binaries.

GnuPG and Git support is the reason why I started using pass and
ultimately switched to gopass. If the latter stops being a viable
password manager, switching back to pass will be much harder with the
new MIME format.

There is also an upstream issue[1] about this and while I haven't read
through *all* the comments, other people seem to have similar reasons
for switching to gopass.

So far however I'm not convinced that my use case will have a future
with gopass, so I'll stay at 1.9.x until I've had the time to properly
research other options or maybe even stay with gopass (and go all-in
with the new MIME format).

[1]: https://github.com/gopasspw/gopass/issues/1365

Signed-off-by: aszlig <aszlig@nix.build>