core/tests: Remove all tests from the "tested" job
The upstream Hydra channels are based on the "tested" job in
release-combined.nix, which already contain a bunch of tests that must
succeed for the channel to advance.

Since evaluation time has skyrocketed for us within the last few months,
we now even don't get *any* evaluation since a while since we hit a
timeout of 10 hours.

I have no idea which changes resulted in this spike of evaluation time
but right now I'm too busy with other things to do a proper
investigation.

To hopefully mitigate this, I decided to base our vuizvui jobset on the
"nixos-unstable" channel and we now only run *additional* tests not
already tested by the upstream channel.

Signed-off-by: aszlig <aszlig@nix.build>