core/tests: Remove runInMachine
The test is not essential for running any real NixOS machines and the
tests just checks whether the runInMachine function for building
derivations inside a VM works.

I was very much reluctant to just remove this test before I actually
fixed the issue upstream because this would mean that it's highly likely
that I'm never going to do it.

However, since all the channels on our side are currently blocking on
this test, which isn't relevant either, I'm going to risk possibly never
fixing runInMachine upstream since I'd have removed that test anyway for
the aforementioned reason.

Signed-off-by: aszlig <aszlig@nix.build>