pkgs/librxtx-java: Fix build with newer JDK
This is actually needed to build the axbo program, which I have rarely
used during the past years. However, I'm not absolutely sure that I
won't be using this again so I decided to fix it, since all that needed
to be done is take an additional patch from Debian that fixes
compatibility with newer JDK versions.

Since the patch changes configure.in and Makefile.am, we need to
regenerate all the files for autotools as well and since there were old
m4 files laying around I deleted them in preAutoreconf.

Signed-off-by: aszlig <aszlig@nix.build>