dnyarri: Set bcache mode to writethrough
I already have this running in writethrough for long enough so that I'm
pretty confident that a disaster like last time[1] should not happen
anytime soon so that hereby I'm making this permanent in the machine
configuration.

The reason why I set this to writearound initially instead of
writethrough, was that I just wanted to go with the most conservative
option first.

Of course, during scrub, we still disable all caching, so we should not
generate too much wear on the SSD.

[1]: 15008e69542774c441e388ad4c2e28a2d27f9ba0

Signed-off-by: aszlig <aszlig@nix.build>