games/starbound: Update to version 1.4.4
I already updated the game a while ago, but didn't commit it yet and
since the update has been out since a year, I think it's time to
actually update it here as well (shame on me).

This essentially includes the "new" bounty hunter update from June 2019,
which has a lot of changes.

So instead of including the changelog here, I'm hereby referring to the
upstream release announcement:

https://playstarbound.com/starbound-1-4-bounty-hunter-update/

I also changed the -Werror flag in the preloader to -Wall, since the
errors/warnings we're getting are false-positives:

  error: '__builtin_strncpy' specified bound depends on the length of
  the source argument [-Werror=stringop-overflow=]

While a warning like this usually is a concern, it's not in our case,
since the size of the destination *also* depends on the length of the
source argument.

Additionally, I changed the test to use the new
networking.interfaces.*.ipv4.addresses option instead of the legacy one
which has been deprecated for ages.

Signed-off-by: aszlig <aszlig@nix.build>