workstation/packages: Remove NixOps
Since the "cryptography" Python library is marked as insecure I'd have
several options here: Add an excemption for the library, switch to
NixOps 2.0 or entirely remove it.

I chose to do the latter, since on all of my deployments I currently use
a heavily patched NixOps version which is somewhat of a mix between 1.5
and 2.0 and the version used here in the workstation profile is unused
since ages anyway.

In case I really need to use nixops on any of my workstations again, I
could still resort to "nix run".

Signed-off-by: aszlig <aszlig@nix.build>