games/gog: Add Graveyard Keeper
This has been mentioned on Discourse[1] and since I have that game in my
GOG library, I just decided to package it.

I only went to the main menu and didn't actually play it, since I'm
somewhat time constrained at the moment. So this only is "for the sake
of completeness" :-)

[1]: https://discourse.nixos.org/t/nix-expression-for-downloaded-gog-game/10595/7

Signed-off-by: aszlig <aszlig@nix.build>