pkgs/psi: Keep urgency hint indefinitely
One thing that has annoyed me since quite a while but ultimately didn't
fix yet, was that the urgency hint was only set for one millisecond.

I don't know how this would look like in a desktop environment, but in
my environment the corresponding workspace only flashes red and then
turns back to blue (the default color) whenever I get a new message.

Since I do not constantly switch to Psi to check whether there is
something new, I sometimes responded very late to messages even though I
didn't want to (eg. not actively working on something).

Of course, I also don't want to be interrupted when I'm actually in
zone, but luckily the way urgency hints are displayed in my environment
is pretty unobtrusive and there is no flashing, blinking or even sounds.

So the only difference is that I do not need to switch to Psi anymore,
to check whether there are new messages.

Signed-off-by: aszlig <aszlig@nix.build>