dnyarri: Fix unit names for btrfs scrub services
Injecting pre/post start scripts into services called
"btrfs-scrub-.service.service" isn't going to do a whole lot if the
actual service name is called "btrfs-scrub-.service".

During the last scrub I was wondering why caching got slow afterwards
and found out that the caching devices were filled with lots of
irrelevant data from the scrub. This led me to inspect what went wrong
and when checking the properties of the scrub service unit, I found out
that it never worked in the first place for the reason meantioned in the
first paragraph.

Actually using the right unit name helps a lot here, so onwards to the
next scrub in February :-)

Signed-off-by: aszlig <aszlig@nix.build>