pkgs/tvl: expose tvl's depot in vuizvui
depot is a nix-based monorepo which contains some great nix utilities
like yants (a nix type system), runTestsuite, mergePatch and so on, a
few interesting pure nix builders like buildLisp and buildGo and a few
packages maintained by @Profpatsch and myself.

This change exposes tvl completely as pkgs.tvl, but prevents hydra from
building it using dontRecurseIntoAttrs as depot pins its own version of
nixpkgs which is not easily overrideable, contains some expensive to
build system configurations we are not interested in and even some
notoriously indeterministic packages.

Additionally it is possible to override pkgs.tvl to use a different or
local version of depot:

    pkgs.tvl.override { tvlSrc = /home/lukas/src/depot; }

To keep with @Profpatsch's previous solution, we pass in vuizvui's
nixpkgs version to depot via nixpkgsBisectPath which may break packages
in depot occasionally if nixpkgs causes breakage in TVL and depot isn't
updated accordingly.