modules/core/common: disable programs.command-not-found by default
programs.command-not-found.enable defaults to true. While it is a neat
feature, it doesn't work with vuizvui because we don't provide a
programs.sqlite database which is required by command-not-found.

Disabling this by default solves the nasty error messages in for example
the default fish configuration if you try to execute a command which is
not installed.

Users which build a custom programs.sqlite database can still enable the
program themselves.