pkgs/sternenseemann/mandoc: init
Build mandoc from cvs and apply a patch which adds an exception for
the checks does makewhatis(8): with this the realpaths of man pages to
be indexed may be symlinks referring to files outside of the root man
directory if they are located in /nix/store.

This allows indexing /run/current-system/sw/share/man using mandoc's
makewhatis(8).

Since fetchcvs is currently broken in nixos-unstable we vendor in the
fixed variant from nixpkgs master until the channel has advanced.