modules/user/sternenseemann: add mandoc module
This module implements a drop-in replacement for documentation.man which
finally lets me get rid of pkgs.man-db. This is still to be considered
experimental as the required patch hasn't landed in upstream mandoc yet.
Should that happen, I'll try to contribute this module back to nixpkgs.

A more detailed description on the module and mandoc on NixOS can be
found at the top of modules/user/sternenseemann/documentation/mandoc.nix