pkgs/profpatsch/nman: if no section is given, treat it as unknown
Previously, we would always assume section 1 was requested if no section
was given on the command line. Now the section is treated as if it was
unknown and we search for a matching man page by looking through the sub
directories of $output/share/man.

This means the following invokation works as expected:

    nman libunwind unw_getcontext

instead of requiring

    nman libunwind 3 unw_getcontext