release.nix: inject programs.sqlite from a nixos channel
programs.command-not-found requires a database containing an index of
all of nixpkgs. Since this is pretty expensive to create we fetch the
database from the nixos-unstable channel. Unfortunately it is not
available via git, so we need to use the tarball which also means hydra
can't automatically update it.

One improvement to this change that would be possible is to also index
vuizvui and merge that with the database from upstream. This should
probably be investigated in the future, but probably affects less people
as the nix tool will only be usable with vuizvui packages as
commond-not-found suggests if vuizvui.modifyNixPath = true.