pkgs/profpatsch/nman: fix panic if mandir.len() < 3
Sometimes share/man contains directories that are not of the form
man<section> which we are checking in build_man_page. However, this
check can't deal with directories that are shorter than 3 characters due
to the use of split_at() which panics if the index is out of range (for
somer reason…).

In nixpkgs it is relatively common to have directories like "de" in
share/man.

This has been resolved.