pkgs/profpatsch/nman: check all man files for matches
Instead of generating n hypothetical filenames for the man page we
search for and checking if one of them exists, we now iterate through
the files in the man dir we are checking and match each of them against
our desired man page and section.

I feel like this makes the code more cumbersome, but on the upside it is
now more unit-testable.