machines/herja: Disable light program
I've been using `illum` to make the screen brightness keys work on this
machine. It's better than invoking light manually or via a xbindrc call.
Although, `illum` is a bit outdated in nixpkgs, and I'm guessing that's
the reason why it crashes when the machine boots. After restarting the
service it's fine. If `illum` isn't updated in nixpkgs soon, I'll have a
look at it.