pkgs/profpatsch: add gpg-private-offline-key
These are the scripts I used to print my gpg private key onte a bunch
of A4 papers, as QR codes of the paperkey output.

It also contains an example of how to automatically read it back in
with a ScanSnap „Einzugsscanner“. I think there also was a test that
checks that the full roundtrip works :)

The QR codes generation was designed in a way that they contain the
highest amount of data when printed on A4 paper, while still being
high-redundancy enough that you can destroy about 1/4th of them before
they become unreadable.

The key was also printed as plain paperkey format, so in the worst
case when I don’t have a scanner I can type it in by hand.