tests/psi: Work around OCR recognition failure
For some reason I currently don't have time to investigate, the test has
stopped to recognise the "Register new account" text, because the text
extracted via OCR was "Reg ster new account" (note: without the "i").

Since we're not testing OCR in general and all we really care here in
the test is whether the application starts up, I just changed the "i" to
a "." (any character in regex) to make sure the detection passes.

Signed-off-by: aszlig <aszlig@nix.build>