machines: Add temporary laptop slylandro
Since Tishtushi had a SSD failure and thus became a majoor nuisance to
work with, I got a temporary laptop from someone (since I don't know
whether they want to be mentioned, I leave out their name for now) in
order to be able to be more productive than waiting for several seconds
for a 1 KiB text file to be saved.

Right now, I'm not sure whether any firmware is needed for the temporary
laptop, so this is a hardware configuration just to get started with a
proper Hydra channel.

Signed-off-by: aszlig <aszlig@nix.build>