factorio: Fix patcher to work with Radare2 5.x
Version 5.0.0 of Radare2 introduced a few changes in symbol naming,
which causes our patcher to fail because it's unable to find the
"/usr/share/factorio" string and the getSystemWriteData() method.

The upstream change[1] in question for example now uses underscores for
certain characters in the symbol/comment name, so for us to be
backwards-compatible I added a check against the major version of
Radare2 to determine whether we should add underscores when needed.

[1]: https://github.com/radareorg/radare2/commit/aaa930ab261a31e58b1257df06db02481cd3bd55

Signed-off-by: aszlig <aszlig@nix.build>