machines/sternenseemann: add bonk shell function
Shared by adisbladis in #nixos-chat this is a very stupid fish function
which bonks a nixpkgs package into the PATH in the current shell. This
is often useful if you need a package ad-hoc and opening a nix-shell is
too much hassle or you find yourself perpetually opening and closing a
nix-shell. Extremely racy with the garbage collector, but this is yolo
ops territory anyways.