fix(profpatsch/write-rust): fail on rust test failure
Port of <https://cl.tvl.fyi/c/depot/+/2657>. Original message:

`forstdin` iterates over the tests in the test directory, and by
default it does *not* fail if an inner loop returns an error, unless
`-o okcodes` is given, a list of exit codes that indicate success.

Now it fails if a loop returns ≠ 0.