modules/profiles/tvl: add tvl profile module
This doesn't do much other than enabling cache.tvl.su as a binary cache
currently, but we can additional settings in the future which are neat
for working with tvl's depot.

I contemplated adding an option to add <depot> to the nixPath, but I
don't want it too desparately right now and it is kind of annoying to
implement for vuizvui, as the default core/common.nix module overrides
any values you set and it is hard to merge the nixPath values, so we'll
probably need an vuizvui.extraNixPath option etc. pp.