modules/core/tests: adjust for new kernel test structure
The kernel-lts and kernel-latest tests seem to have been removed (not
sure when exactly, didn't blame). Instead we can use
nixosTests.kernel-generic."linux_${major}_${minor}". This allows us to
easily check for tests for arbitrary kernel versions as well.

However there is currently a bug that these are returned as functions
instead of derivations at the moment which will hopefully be fixed by my
upstream PR soon. Until that happens we check if the test is a
derivation or not with lib.isDerivation.

cc @aszlig for comments / further improvements. Main intention right now
is to fix eval in a future proof way.