machines/herja: Add pulseaudio for pactl
While I'm generally happy with `pipewire`, some things broke, such as
volume control. I've tried to fix it using a range of front-ends for PA,
but so far only `pactl` seems to be behaving correctly (but broke mic
mute again). I might have to switch back to PA in the end as it's also
broken my setup on gunnr somewhat.