machines/profpatsch: make bitlbee and philip normal users
https://github.com/NixOS/nixpkgs/pull/115332 introduced a new check that
requires either isSystemUser or isNormalUser to be true to help UID
collision checks (I don't remember how this exactly aids that). To fix
evaluation of @Profpatsch's machines, I've set isNormalUser for bitlbee
and philip.

Note that bitlbee probably should be a system user, but seems like there
may be issues with that: https://github.com/NixOS/nixpkgs/pull/115332#issuecomment-792287827
Just to be safe, leave it as a normal user for now, the proper fix seems
to be to move the data dir creation into the service setup, rather than
doing it via the home setup.