pkgs/psi: Replaced aliased use of html-tidy
A recent change[1] disabled aliases by default in VM tests and since
libtidy actually has been an alias of html-tidy since 2014 it's a good
idea to use the actual non-aliased packaged.

Since I added my PSI build in 2019, I probably didn't check for whether
the package name in nixpkgs would be different while packaging and only
used the name as reported by CMake, thinking it would work (which it
did).

Disallowing aliases in VM tests however is a good change, so let's use
the real package name.

This should fix the evaluation of the Hydra jobset.

[1]: https://github.com/NixOS/nixpkgs/commit/3edde6562e19698da69a499881e0a2e4f5a497a2

Signed-off-by: aszlig <aszlig@nix.build>