psi: Update to latest upstream master
This is mainly to incorporate my latest fix[1] for OMEMO, so in theory
updating the plugins would have been sufficient. However, since I like
to eat the freshest set of new bugs, I also updated everything else
except the theme. The latter seems to be a bit more complicated, since
it changed the way they're building it so I skipped that for now.

[1]: https://github.com/psi-im/plugins/pull/91

Signed-off-by: aszlig <aszlig@nix.build>