pkgs/profpatsch: Replace s6PortableUtils alias
Similar to 4701a995cb865c5d7178f574a3eae5872595e768, where I replaced
the libtidy alias for html-tidy because it broke evaluation of the PSI
test, I found another test for nman which uses an alias.

The background is that aliases are now[1] no longer allowed in NixOS VM
tests and since "s6PortableUtils" is indirectly referenced, we get an
evaluation error on Hydra.

Using the unaliased name fixes evaluation and should not change anything
in functionality.

[1]: https://github.com/NixOS/nixpkgs/commit/3edde6562e19698da69a499881e0a2e4f5a497a2

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch