treewide: Replace pkgconfig with pkg-config
This is another alias which got introduced in 2018, because the actual
command is "pkg-config" and so the package name containing a dash is
more reasonable.

The reason why I'm doing this is because NixOS VM tests now disallow
aliases and while the evaluation error in question only affected the
"gnupg" test, I decided to change all occurences in the event that we
might want to disallow aliases for things other than VM tests.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @sternenseemann for "opam-env"